requirements.txt
Flask==2.0.1
mysql-connector-python
mysql-connector~=2.2.9
Flask-Session==0.4.0
flask-talisman~=0.7.0
Flask==2.0.1
pandas==1.3.4


